+++ src/geoai3d/__init__.py
@@ -19,6 +19,7 @@ from geoai3d.io.xyz import read_xyz, to_xyz
 from geoai3d.outofcore import estimate_radius, geometric_features_tiled
 from geoai3d.stream import geometric_features_stream
 from geoai3d.subsample import subsample
+from geoai3d.viz import view
 
 __version__ = "0.0.1"
 
@@ -44,4 +45,5 @@ __all__ = [
     "to_las",
     "to_parquet",
     "to_xyz",
+    "view",
 ]